[FEATURE] Title tag rendering and field to override content
The title tag is rendered as a reverse breadcrumb now by default. If the
new "Page title override" field is filled in, it's content will be used
instead.

The new field is delivered as a mixin node type and can be pulled in to
custom types easily (TYPO3.Neos.Seo:TitleTagMixin).

Changing the rendering can be adjusted via `head.titleTag.default` and
`head.titleTag.content` on the `TYPO3.Neos:Page` TypoScript object as
needed.

Change-Id: Ie0410dfc9a3ec37c11aa7360f4eccb7b9877a345
Releases: master
Resolves: NEOS-707