[FEATURE] Open Graph Support
The Open Graph protocol enables any web page to become a rich object
in a social graph by adding custom meta tags to the page.

The essential ones are:
* og:title
* og:type
* og:url
* og:image
* og:description

In general ``og:`` tags are only shown if they have data, because
otherwise Facebook for example will extract data for the generated view
from the page itself. If you are not satisfied with the generated output
you have the possibility to customize it to your needs using TypoScript.

In general they are only shown if a ``og:type`` is selected.

* ``og:title`` has no fallback
* ``og:description`` will use as fallback ``meta:description``
* ``og:url`` contains the url of the document
* ``og:image`` has no fallback

Change-Id: I41ad6bfb31907c7f964a5fbe09ff39058b84ce13
Releases: master, 1.0
Resolves: NEOS-718