BUGFIX: hreflang attribute should contain valid HTML language tags
The alternate language links and the html lang attribute contain
the configured values for the ``language`` dimension if it exists.
We assume a valid locale to be separated by underscores as in
``is_IS`` for icelandic spoken on iceland. HTML expects this
separated with dashes though so we need to convert this on output.

NEOS-1344 #close