BUGFIX: Exclude shortcuts from XML sitemap
When shortcuts are included in the XML sitemap, that may lead to
invalid sitemap entries: If a shortcut points to an external URL,
it violates the XML sitemap protocol.

Setting the filter property of the used Menu object to exclude Shortcuts
does not work as expected, as it would exclude children of those nodes
as well.

Thus a check in the XML template is used.