!!! FEATURE: Make titletag configurable and always append suffix
This will allow configuring a generic suffix for the title tag.
Usually this would be the sites name or something similar.

For specific purposes the title generation can be extended
by providing additional cases in Fusion.

This is breaking as the sitenode will now render it’s own title too,
and previous overrides on the title tag might behave differently.

The title of the sitenode will only appear when visiting the site node but
not on subpages in the titles breadcrumbs.

Resolves: #81