BUGFIX: Invalidate sitemap caches properly
The xml-sitemap  did have cache-tags but no cache-mode and cache-identifier. 
Since the default cache mode is embed the sitemap cache was embedded into the document 
cache for the document and thus only updated when the siteNode was changed.

This change adds the cache-identifier and the cache-mode. It also puts the startingPoint into the context as
somehow it was not available in the entry identifier otherwise.

resolves: #141